requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
discord.py @ git+https://github.com/Rapptz/discord.py@37076a04de36031d0c39664cafe0a321aa9595a3
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
python-a2s==1.3.0
yarl==1.7.2
